Split test_products_data file into smaller test files
