+++ saleor/csv/tests/export/products_data/test_handle_relations_data.py
@@ -1,269 +1,19 @@
 from unittest.mock import patch
 
-from measurement.measures import Weight
-
-from ....graphql.csv.enums import ProductFieldEnum
-from ....product.models import (
-    Attribute,
-    Product,
-    ProductImage,
-    ProductVariant,
-    VariantImage,
-)
-from ....warehouse.models import Warehouse
-from ...utils.products_data import (
+from .....graphql.csv.enums import ProductFieldEnum
+from .....product.models import Attribute, Product, ProductImage, VariantImage
+from .....warehouse.models import Warehouse
+from ....utils.products_data import (
     ProductExportFields,
     add_attribute_info_to_data,
     add_collection_info_to_data,
     add_image_uris_to_data,
     add_warehouse_info_to_data,
-    get_attributes_headers,
-    get_export_fields_and_headers_info,
-    get_product_export_fields_and_headers,
-    get_products_data,
     get_products_relations_data,
     get_variants_relations_data,
-    get_warehouses_headers,
     prepare_products_relations_data,
     prepare_variants_relations_data,
 )
-
-
-def test_get_products_data(product, product_with_image, collection, image):
-    # given
-    product.weight = Weight(kg=5)
-    product.save()
-
-    collection.products.add(product)
-
-    variant = product.variants.first()
-    VariantImage.objects.create(variant=variant, image=product.images.first())
-
-    products = Product.objects.all()
-    export_fields = set(
-        value
-        for mapping in ProductExportFields.HEADERS_TO_FIELDS_MAPPING.values()
-        for value in mapping.values()
-    )
-    warehouse_ids = [str(warehouse.pk) for warehouse in Warehouse.objects.all()]
-    attribute_ids = [str(attr.pk) for attr in Attribute.objects.all()]
-
-    variants = []
-    for variant in product.variants.all():
-        for attr in variant.attributes.all():
-            attribute_ids.append(str(attr.assignment.attribute.pk))
-        variant.weight = Weight(kg=3)
-        variants.append(variant)
-
-    ProductVariant.objects.bulk_update(variants, ["weight"])
-
-    variants = []
-    for variant in product_with_image.variants.all():
-        variant.weight = None
-        variants.append(variant)
-    ProductVariant.objects.bulk_update(variants, ["weight"])
-
-    # when
-    result_data = get_products_data(
-        products, export_fields, attribute_ids, warehouse_ids
-    )
-
-    # then
-    expected_data = []
-    for product in products.order_by("pk"):
-        product_data = {
-            "id": product.id,
-            "name": product.name,
-            "is_published": product.is_published,
-            "description": product.description,
-            "category__slug": product.category.slug,
-            "price_amount": product.price_amount,
-            "currency": product.currency,
-            "product_type__name": product.product_type.name,
-            "charge_taxes": product.charge_taxes,
-            "collections__slug": (
-                ""
-                if not product.collections.all()
-                else product.collections.first().slug
-            ),
-            "product_weight": (
-                "{} g".format(int(product.weight.value * 1000))
-                if product.weight
-                else ""
-            ),
-            "images__image": (
-                ""
-                if not product.images.all()
-                else "http://mirumee.com{}".format(product.images.first().image.url)
-            ),
-        }
-
-        assigned_attribute = product.attributes.first()
-        if assigned_attribute:
-            header = f"{assigned_attribute.attribute.slug} (product attribute)"
-            product_data[header] = assigned_attribute.values.first().slug
-
-        for variant in product.variants.all():
-            data = {
-                "variants__sku": variant.sku,
-                "variants__currency": variant.currency,
-                "variants__price_override_amount": variant.price_override_amount,
-                "variants__cost_price_amount": variant.cost_price_amount,
-                "variants__images__image": (
-                    ""
-                    if not variant.images.all()
-                    else "http://mirumee.com{}".format(variant.images.first().image.url)
-                ),
-                "variant_weight": (
-                    "{} g".foramt(int(variant.weight.value * 1000))
-                    if variant.weight
-                    else ""
-                ),
-            }
-            data.update(product_data)
-
-            for stock in variant.stocks.all():
-                slug = stock.warehouse.slug
-                warehouse_headers = [
-                    f"{slug} (warehouse quantity)",
-                ]
-                data[warehouse_headers[0]] = stock.quantity
-
-            assigned_attribute = variant.attributes.first()
-            if assigned_attribute:
-                header = f"{assigned_attribute.attribute.slug} (variant attribute)"
-                data[header] = assigned_attribute.values.first().slug
-
-            expected_data.append(data)
-
-    assert result_data == expected_data
-
-
-def test_get_products_data_for_specified_attributes(
-    product, product_with_variant_with_two_attributes
-):
-    # given
-    products = Product.objects.all()
-    export_fields = {"id", "variants__sku"}
-    attribute_ids = [str(attr.pk) for attr in Attribute.objects.all()][:1]
-
-    # when
-    result_data = get_products_data(products, export_fields, attribute_ids, [])
-
-    # then
-    expected_data = []
-    for product in products.order_by("pk"):
-        product_data = {"id": product.pk}
-
-        for assigned_attribute in product.attributes.all():
-            if assigned_attribute:
-                header = f"{assigned_attribute.attribute.slug} (product attribute)"
-                if str(assigned_attribute.attribute.pk) in attribute_ids:
-                    product_data[header] = assigned_attribute.values.first().slug
-
-        for variant in product.variants.all():
-            data = {}
-            data.update(product_data)
-            data["variants__sku"] = variant.sku
-            for assigned_attribute in variant.attributes.all():
-                header = f"{assigned_attribute.attribute.slug} (variant attribute)"
-                if str(assigned_attribute.attribute.pk) in attribute_ids:
-                    data[header] = assigned_attribute.values.first().slug
-
-            expected_data.append(data)
-
-    assert result_data == expected_data
-
-
-def test_get_products_data_for_specified_warehouses(
-    product, product_with_image, variant_with_many_stocks
-):
-    # given
-    product.variants.add(variant_with_many_stocks)
-
-    products = Product.objects.all()
-    export_fields = {"id", "variants__sku"}
-    warehouse_ids = [str(warehouse.pk) for warehouse in Warehouse.objects.all()][:2]
-    attribute_ids = []
-
-    # when
-    result_data = get_products_data(
-        products, export_fields, attribute_ids, warehouse_ids
-    )
-
-    # then
-    expected_data = []
-    for product in products.order_by("pk"):
-        product_data = {"id": product.pk}
-
-        for variant in product.variants.all():
-            data = {"variants__sku": variant.sku}
-            data.update(product_data)
-
-            for stock in variant.stocks.all():
-                if str(stock.warehouse.pk) in warehouse_ids:
-                    slug = stock.warehouse.slug
-                    warehouse_headers = [
-                        f"{slug} (warehouse quantity)",
-                    ]
-                    data[warehouse_headers[0]] = stock.quantity
-
-            expected_data.append(data)
-
-    for res in result_data:
-        assert res in expected_data
-
-
-def test_get_products_data_for_specified_warehouses_and_attributes(
-    product,
-    variant_with_many_stocks,
-    product_with_image,
-    product_with_variant_with_two_attributes,
-):
-    # given
-    product.variants.add(variant_with_many_stocks)
-
-    products = Product.objects.all()
-    export_fields = {"id", "variants__sku"}
-    warehouse_ids = [str(warehouse.pk) for warehouse in Warehouse.objects.all()]
-    attribute_ids = [str(attr.pk) for attr in Attribute.objects.all()]
-
-    # when
-    result_data = get_products_data(
-        products, export_fields, attribute_ids, warehouse_ids
-    )
-
-    # then
-    expected_data = []
-    for product in products.order_by("pk"):
-        product_data = {"id": product.id}
-
-        for assigned_attribute in product.attributes.all():
-            if assigned_attribute:
-                header = f"{assigned_attribute.attribute.slug} (product attribute)"
-                if str(assigned_attribute.attribute.pk) in attribute_ids:
-                    product_data[header] = assigned_attribute.values.first().slug
-
-        for variant in product.variants.all():
-            data = {"variants__sku": variant.sku}
-            data.update(product_data)
-
-            for stock in variant.stocks.all():
-                if str(stock.warehouse.pk) in warehouse_ids:
-                    slug = stock.warehouse.slug
-                    warehouse_headers = [
-                        f"{slug} (warehouse quantity)",
-                    ]
-                    data[warehouse_headers[0]] = stock.quantity
-
-            for assigned_attribute in variant.attributes.all():
-                header = f"{assigned_attribute.attribute.slug} (variant attribute)"
-                if str(assigned_attribute.attribute.pk) in attribute_ids:
-                    data[header] = assigned_attribute.values.first().slug
-
-            expected_data.append(data)
-
-    assert result_data == expected_data
 
 
 @patch("saleor.csv.utils.products_data.prepare_products_relations_data")
@@ -831,267 +581,3 @@ def test_add_warehouse_info_to_data_data_no_slug(product):
 
     # then
     assert result == input_data
-
-
-def test_get_export_fields_and_headers_fields_with_price():
-    # given
-    export_info = {
-        "fields": [
-            ProductFieldEnum.PRICE.value,
-            ProductFieldEnum.PRICE_OVERRIDE.value,
-            ProductFieldEnum.COLLECTIONS.value,
-            ProductFieldEnum.DESCRIPTION.value,
-        ],
-        "warehoses": [],
-    }
-
-    # when
-    export_fields, file_headers = get_product_export_fields_and_headers(export_info)
-
-    # then
-    expected_headers = [
-        "id",
-        "price",
-        "product currency",
-        "price override",
-        "variant currency",
-        "collections",
-        "description",
-    ]
-
-    expected_fields = [
-        "id",
-        "price_amount",
-        "currency",
-        "variants__price_override_amount",
-        "variants__currency",
-        "collections__slug",
-        "description",
-    ]
-
-    assert export_fields == expected_fields
-    assert file_headers == expected_headers
-
-
-def test_get_export_fields_and_headers_fields_with_cost_price():
-    # given
-    export_info = {
-        "fields": [
-            ProductFieldEnum.NAME.value,
-            ProductFieldEnum.COST_PRICE.value,
-            ProductFieldEnum.COLLECTIONS.value,
-        ],
-        "warehoses": [],
-    }
-
-    # when
-    export_fields, file_headers = get_product_export_fields_and_headers(export_info)
-
-    # then
-    expected_headers = [
-        "id",
-        "name",
-        "cost price",
-        "variant currency",
-        "collections",
-    ]
-
-    expected_fields = [
-        "id",
-        "name",
-        "variants__cost_price_amount",
-        "variants__currency",
-        "collections__slug",
-    ]
-
-    assert export_fields == expected_fields
-    assert file_headers == expected_headers
-
-
-def test_get_export_fields_and_headers_fields_with_cost_price_and_price_override():
-    # given
-    export_info = {
-        "fields": [
-            ProductFieldEnum.NAME.value,
-            ProductFieldEnum.PRICE_OVERRIDE.value,
-            ProductFieldEnum.COST_PRICE.value,
-            ProductFieldEnum.COLLECTIONS.value,
-        ],
-        "warehoses": [],
-    }
-
-    # when
-    export_fields, file_headers = get_product_export_fields_and_headers(export_info)
-
-    # then
-    expected_headers = [
-        "id",
-        "name",
-        "price override",
-        "variant currency",
-        "cost price",
-        "collections",
-    ]
-
-    expected_fields = [
-        "id",
-        "name",
-        "variants__price_override_amount",
-        "variants__currency",
-        "variants__cost_price_amount",
-        "collections__slug",
-    ]
-
-    assert export_fields == expected_fields
-    assert file_headers == expected_headers
-
-
-def test_get_export_fields_and_headers_fields_without_price():
-    # given
-    export_info = {
-        "fields": [
-            ProductFieldEnum.COLLECTIONS.value,
-            ProductFieldEnum.DESCRIPTION.value,
-            ProductFieldEnum.VARIANT_SKU.value,
-        ],
-        "warehoses": [],
-    }
-
-    # when
-    export_fields, file_headers = get_product_export_fields_and_headers(export_info)
-
-    # then
-    expected_headers = ["id", "collections", "description", "variant sku"]
-
-    assert set(export_fields) == {
-        "collections__slug",
-        "id",
-        "variants__sku",
-        "description",
-    }
-    assert file_headers == expected_headers
-
-
-def test_get_export_fields_and_headers_no_fields():
-    export_fields, file_headers = get_product_export_fields_and_headers({})
-
-    assert export_fields == ["id"]
-    assert file_headers == ["id"]
-
-
-def test_get_attributes_headers(product_with_multiple_values_attributes):
-    # given
-    attribute_ids = Attribute.objects.values_list("id", flat=True)
-    export_info = {"attributes": attribute_ids}
-
-    # when
-    attributes_headers = get_attributes_headers(export_info)
-
-    # then
-    product_headers = []
-    variant_headers = []
-    for attr in Attribute.objects.all():
-        if attr.product_types.exists():
-            product_headers.append(f"{attr.slug} (product attribute)")
-        if attr.product_variant_types.exists():
-            variant_headers.append(f"{attr.slug} (variant attribute)")
-
-    expected_headers = product_headers + variant_headers
-    assert attributes_headers == expected_headers
-
-
-def test_get_attributes_headers_lack_of_attributes_ids():
-    # given
-    export_info = {}
-
-    # when
-    attributes_headers = get_attributes_headers(export_info)
-
-    # then
-    assert attributes_headers == []
-
-
-def test_get_warehouses_headers(warehouses):
-    # given
-    warehouse_ids = [warehouses[0].pk]
-    export_info = {"warehouses": warehouse_ids}
-
-    # when
-    warehouse_headers = get_warehouses_headers(export_info)
-
-    # then
-    assert warehouse_headers == [f"{warehouses[0].slug} (warehouse quantity)"]
-
-
-def test_get_warehouses_headers_lack_of_warehouse_ids():
-    # given
-    export_info = {}
-
-    # when
-    warehouse_headers = get_warehouses_headers(export_info)
-
-    # then
-    assert warehouse_headers == []
-
-
-def test_get_export_fields_and_headers_info(
-    warehouses, product_with_multiple_values_attributes
-):
-    # given
-    warehouse_ids = [w.pk for w in warehouses]
-    attribute_ids = [attr.pk for attr in Attribute.objects.all()]
-    export_info = {
-        "fields": [
-            ProductFieldEnum.PRICE.value,
-            ProductFieldEnum.PRICE_OVERRIDE.value,
-            ProductFieldEnum.COLLECTIONS.value,
-            ProductFieldEnum.DESCRIPTION.value,
-        ],
-        "warehouses": warehouse_ids,
-        "attributes": attribute_ids,
-    }
-
-    expected_file_headers = [
-        "id",
-        "price",
-        "product currency",
-        "price override",
-        "variant currency",
-        "collections",
-        "description",
-    ]
-
-    # when
-    export_fields, file_headers, data_headers = get_export_fields_and_headers_info(
-        export_info
-    )
-
-    # then
-    expected_fields = [
-        "id",
-        "price_amount",
-        "currency",
-        "variants__price_override_amount",
-        "variants__currency",
-        "collections__slug",
-        "description",
-    ]
-
-    product_headers = []
-    variant_headers = []
-    for attr in Attribute.objects.all().order_by("slug"):
-        if attr.product_types.exists():
-            product_headers.append(f"{attr.slug} (product attribute)")
-        if attr.product_variant_types.exists():
-            variant_headers.append(f"{attr.slug} (variant attribute)")
-
-    warehouse_headers = [f"{w.slug} (warehouse quantity)" for w in warehouses]
-    excepted_headers = (
-        expected_fields + product_headers + variant_headers + warehouse_headers
-    )
-
-    expected_file_headers += product_headers + variant_headers + warehouse_headers
-
-    assert expected_file_headers == file_headers
-    assert set(export_fields) == set(expected_fields)
-    assert data_headers == excepted_headers
